requirements.txt
asgiref==3.6.0 ; python_version >= "3.10" and python_version < "4.0"
certifi==2022.12.7 ; python_version >= "3.10" and python_version < "4"
charset-normalizer==3.0.1 ; python_version >= "3.10" and python_version < "4"
crispy-bootstrap5==0.7 ; python_version >= "3.10" and python_version < "4.0"
defusedxml==0.7.1 ; python_version >= "3.10" and python_version < "4.0"
diff-match-patch==20200713 ; python_version >= "3.10" and python_version < "4.0"
dj-database-url==1.2.0 ; python_version >= "3.10" and python_version < "4.0"
django-ckeditor==6.5.1 ; python_version >= "3.10" and python_version < "4.0"
django-crispy-forms==2.0 ; python_version >= "3.10" and python_version < "4.0"
django-extensions==3.2.1 ; python_version >= "3.10" and python_version < "4.0"
django-import-export==3.0.2 ; python_version >= "3.10" and python_version < "4.0"
django-js-asset==2.0.0 ; python_version >= "3.10" and python_version < "4.0"
django-model-utils==4.3.1 ; python_version >= "3.10" and python_version < "4.0"
django-ordered-model==3.6 ; python_version >= "3.10" and python_version < "4.0"
django==4.1.6 ; python_version >= "3.10" and python_version < "4.0"
et-xmlfile==1.1.0 ; python_version >= "3.10" and python_version < "4.0"
gunicorn==20.1.0 ; python_version >= "3.10" and python_version < "4.0"
idna==3.4 ; python_version >= "3.10" and python_version < "4"
markuppy==1.14 ; python_version >= "3.10" and python_version < "4.0"
newrelic==8.6.0 ; python_version >= "3.10" and python_version < "4.0"
odfpy==1.4.1 ; python_version >= "3.10" and python_version < "4.0"
openpyxl==3.1.1 ; python_version >= "3.10" and python_version < "4.0"
pillow==9.4.0 ; python_version >= "3.10" and python_version < "4.0"
psycopg2==2.9.5 ; python_version >= "3.10" and python_version < "4.0"
python-decouple==3.7 ; python_version >= "3.10" and python_version < "4.0"
pyyaml==6.0 ; python_version >= "3.10" and python_version < "4.0"
requests==2.28.2 ; python_version >= "3.10" and python_version < "4"
setuptools==67.1.0 ; python_version >= "3.10" and python_version < "4.0"
sqlparse==0.4.3 ; python_version >= "3.10" and python_version < "4.0"
tablib[html,ods,xls,xlsx,yaml]==3.3.0 ; python_version >= "3.10" and python_version < "4.0"
tzdata==2022.7 ; python_version >= "3.10" and python_version < "4.0" and sys_platform == "win32"
urllib3==1.26.14 ; python_version >= "3.10" and python_version < "4"
xlrd==2.0.1 ; python_version >= "3.10" and python_version < "4.0"
xlwt==1.3.0 ; python_version >= "3.10" and python_version < "4.0"
